clk: clk-xlnx-clock-wizard-v: fix wrong handling of 64bit divisions on 32bit ARM
On a 32bit ILP32 operating system (Linux on 32bit x86 and ARM) K&R C,
C89 and C90 plain number defines default to int (aka int32_t) and to
unsigned long long (aka uint64_t) in C99 and beyond.

The constants WZRD_VCO_MIN and WZRD_VCO_MAX also exceed the range of int
and unsigned int (2160000000 > 2147483648 and 4320000000 > 4294967296).
They need to be defined as 64bit. The additional constants WZRD_M_MIN,
WZRD_M_MAX, WZRD_D_MIN, WZRD_D_MAX, WZRD_O_MIN and WZRD_O_MAX and the
variables in clk_wzrd_get_divisors() are all involved in singned 64bit
additions, substractions and divisions (often promoted to modulo ops).
Because of that and to prevent type mismatches, the constants need to be
signed and the *VCO* constants need to be 64bit. The loop variables need
to be signed 32bit (because of the usage as divisors) and the variables
dealing with the rate need to be 64bit (dividend).

That in turn results in 64bit signed divisions/modulo, which are
basically not implemented in hardware on any ARMv7 and are not provided
as arch depended kernel library code. In that case math64.h macros need
to be used to issue proper 64bit ARMv7 machine code. Even worse, there
is only support for 64bit dividends and 32bit divisors in the kernel
math library, if signed values are used on ARMv7. In the case of 64bit
ARMv8 the math64 macros use different code paths, so the same macros
work nicely with 32 and 64 bit architectures.

drivers/clk/clk-xlnx-clock-wizard-v.c: In function 'clk_wzrd_get_divisors':
drivers/clk/clk-xlnx-clock-wizard-v.c:230:25: warning: this decimal constant is unsigned only in ISO C90
  230 |                         if (vco_freq >= WZRD_VCO_MIN && vco_freq <= WZRD_VCO_MAX) {
      |                         ^~

ERROR: modpost: "__aeabi_ldivmod" [drivers/clk/clk-xlnx-clock-wizard-v.ko] undefined!
ERROR: modpost: "__aeabi_uldivmod" [drivers/clk/clk-xlnx-clock-wizard-v.ko] undefined!

Fixes: 8b5383767cb4 ("clocking-wizard: Add versal clocking wizard support")
Signed-off-by: Wilken Gottwalt <12194808+wgottwalt@users.noreply.github.com>